refactor: change the type of timestampMillisec to long (#1086)

+++ Assets/MediaPipeUnity/Samples/Scenes/Tasks/Face Landmark Detection/FaceLandmarkerRunner.cs
@@ -110,7 +110,7 @@ namespace Mediapipe.Unity.Sample.FaceLandmarkDetection
       }
     }
 
-    private void OnFaceLandmarkDetectionOutput(FaceLandmarkerResult result, Image image, int timestamp)
+    private void OnFaceLandmarkDetectionOutput(FaceLandmarkerResult result, Image image, long timestamp)
     {
       _faceLandmarkerResultAnnotationController.DrawLater(result);
     }
